fix: support RuneLite 1.13 long GE prices and >2.147b trades (max cash) (#277)
* fix: compile against RuneLite 1.13 long GE price API and widen GP money to long

RuneLite 1.13 (api: update 241) widens GrandExchangeOffer.getPrice()/getSpent()
and ItemManager.getItemPrice() to long ahead of Jagex's Sep 30 max-cash update,
after which a single GE price can exceed Integer.MAX_VALUE. master failed to
compile against 1.13.0-SNAPSHOT with 10 lossy long->int conversions.

Every receiving local, field and parameter is widened to long instead of cast,
so the plugin compiles against both latest.release (1.12, int widens
implicitly) and latest.integration (1.13) and can target plugin-hub master.
Nothing narrows a GE price or spent back to int.

The same sweep widens every GP value that would break above 2.147b (#1364):
- Gson response and request DTOs (FlipRecommendation, ActiveFlip, FlipAnalysis,
  FavoriteItem, OfferAdvice*, FlipAdjustment*, SellPriceCheck*, Transaction/
  HistoryBackfill/BankItem requests, FlipFinderResponse.cashStack). An int field
  threw JsonSyntaxException and dropped the whole response.
- WikiPrice parsing uses getAsLong.
- Silent overflow/saturation: slot tooltip P&L, Flip Assist total profit,
  FocusedFlip adjusted profit, GeTax breakeven, SmartSellPricer min profitable
  (+1 wrapped negative), Flip Finder 5% markup, SessionStats gp/hr.
- GeTax.breakevenSellPrice starts from min(closed form, buy + cap), so a capped
  price no longer walks down millions of steps.
- GpUtils.parseDigits and GEHistoryService parsing return long; GE History rows
  above int32 are no longer dropped.
- Formatter/util signatures, session/adjustment/stale-offer price maps.
- Persisted state: OfferRecord.price, AutoRecommendService buyPrices,
  ExitTrades buyBasis. Old int JSON still loads into the long fields; a
  downgrade after writing >int32 values loses that state (restores already
  catch the exception).

Left int by design: quantities, item ids, slots, GE tax (capped at 5m), the
coin-stack cash stack and inventory GP (cash stack cap unchanged), config
thresholds, and the GE_NEWOFFER_PRICE varbit read (a varbit cannot carry more
than 32 bits; negative reads already clamp to 0).

Tests stub RuneLite money getters through GpAnswer so the suite compiles and
passes on both 1.12 and 1.13.

Refs Flip-Smart/flip-smart#1363, Flip-Smart/flip-smart#1364

* test: cover >2.147b GE prices across parse, tax, profit and persistence

Pins the max-cash widening with a 3,000,000,000 gp per-item price and ~2.1T
totals: Gson parse of FlipRecommendation and the other money DTOs, wiki price
parse, GeTax cap/breakeven (including a bound on the breakeven walk),
SmartSellPricer min profitable price, the profit calc sites (adjusted profit,
slot tooltip P&L, card metrics, realized profit, sell description, gp/hr),
GpUtils.parseDigits and GE History text parse, and old int-valued persisted
JSON (OfferRecord, AutoRecommend and Exit Trades state) loading into the
widened long fields.

Refs Flip-Smart/flip-smart#1364

* fix: add B/T GP formatting tiers and cap-aware min profitable sell price

Above 1b the compact GP formatters only had an M tier, so 66.5b rendered as
"66500.0M" and 2.1T as "2100000.0M". GpUtils.formatGP/formatGPSigned (and
PanelFormat.formatGP through them) and the offer-description shorthand now
step through K/M/B/T with the M tier's existing one-decimal format.

SmartSellPricer.calculateMinProfitableSellPrice used buy / 0.98 + 1, ignoring
the 5m per-item GE tax cap, so above ~250m it overstated the minimum (3.061b
instead of 3.005b for a 3b buy). It now bounds the estimate by buy + tax cap,
leaving results below the cap unchanged.

Refs Flip-Smart/flip-smart#1364

* fix(quality): resolve codacy PMD_field-method-name-collision in Dtos.java

Rename TransactionRequest.Builder's private recommendedSellPrice field to
recommendedSellPriceValue so it no longer shadows the fluent setter method
of the same name. Public API (the recommendedSellPrice(Long) builder
method and the outer class's recommendedSellPrice field) is unchanged.

* chore: trigger Codacy re-analysis after dismissing V9 param-count noise

Lizard_parameter-count-medium and PMD_ExcessiveParameterList on
buildReadjustmentBody (10 params) predate this branch's diff — the
method and its full parameter list came from V9 (#276) on master, not
from this PR's int->long widening. Dismissed via Codacy API rather
than refactored to stay token-neutral against the 194k Hub ceiling.


---------

+++ src/main/java/com/flipsmart/AutoRecommendService.java
@@ -202,7 +202,7 @@ public class AutoRecommendService
 
 	// Provider for the panel's displayed (smart) sell price — preferred over session's stored price
 	@Setter
-	private volatile IntFunction<Integer> displayedSellPriceProvider;
+	private volatile IntFunction<Long> displayedSellPriceProvider;
 
 	// Last overlay message sent — readable by the overlay as a fallback when the
 	// async callback result gets lost due to race conditions
@@ -241,7 +241,7 @@ public class AutoRecommendService
 
 		// Buy prices stored when buy orders were placed — used as cost basis for sell adjustments.
 		// Persisted because backend doesn't track the exact price the user paid.
-		Map<Integer, Integer> buyPrices;
+		Map<Integer, Long> buyPrices;
 	}
 
 	public AutoRecommendService(FlipSmartConfig config, FlipSmartPlugin plugin, OfferStore offerStore)
@@ -502,7 +502,7 @@ public class AutoRecommendService
 		// Don't re-show sell overlay if this item already has an active sell order — unless the
 		// advisor has a pending reprice, in which case surface the new price on the setup screen
 		// immediately instead of blanking until the old offer clears.
-		Integer resellPrice = staleOffers.getResellPrice(itemId);
+		Long resellPrice = staleOffers.getResellPrice(itemId);
 		boolean repricePending = resellPrice != null && resellPrice > 0;
 		if (offerStore.hasActiveSellOfferForItem(itemId) && !repricePending)
 		{
@@ -511,7 +511,7 @@ public class AutoRecommendService
 			// returning blank — the price lives in session state, so there's no need to wait for
 			// a later event/tick. Off-screen we still bail, to avoid re-surfacing a listed sell.
 			boolean setupScreenOpen = java.util.Objects.equals(queue.getLockedItemId(), itemId);
-			Integer knownPrice = resolveBestSellPrice(itemId);
+			Long knownPrice = resolveBestSellPrice(itemId);
 			if (setupScreenOpen && knownPrice != null && knownPrice > 0)
 			{
 				int qty = resolveRepriceQuantity(itemId);
@@ -535,7 +535,7 @@ public class AutoRecommendService
 			return SellFocusResult.FOCUSED;
 		}
 
-		Integer sellPrice = resolveBestSellPrice(itemId);
+		Long sellPrice = resolveBestSellPrice(itemId);
 		if (sellPrice == null || sellPrice <= 0)
 		{
 			// Try to recover sell price from recommendation queue
@@ -786,7 +786,7 @@ public class AutoRecommendService
 			// directly, mirroring handleBuyCollected — instead of advancing to a new buy for the
 			// freed slot (where the resolver's S2 empty-slot buy would outrank the S3 re-list).
 			PlayerSession session = plugin.getSession();
-			Integer sellPrice = session != null ? session.getRecommendedPrice(itemId) : null;
+			Long sellPrice = session != null ? session.getRecommendedPrice(itemId) : null;
 			if (session != null && session.getCollectedItemIds().contains(itemId) && sellPrice != null && sellPrice > 0)
 			{
 				log.debug("Auto-recommend: Sell modified/returned for {} - re-listing", itemName);
@@ -816,7 +816,7 @@ public class AutoRecommendService
 
 		ensureSellPriceAvailable(itemId);
 		boolean isCollected = session.getCollectedItemIds().contains(itemId);
-		Integer sellPrice = session.getRecommendedPrice(itemId);
+		Long sellPrice = session.getRecommendedPrice(itemId);
 		// Collecting must not downgrade a partial-cancel (S1 list) to a completed-buy (S3),
 		// or the resolver lets an empty-slot buy (S2) outrank listing the held items.
 		CollectOrigin priorOrigin = session.getCollectOrigin(itemId);
@@ -1372,7 +1372,7 @@ public class AutoRecommendService
 	 * When the timer expires, checkAdjustmentTimers() will prompt the user
 	 * to adjust the price if the recommendation has changed.
 	 */
-	private void scheduleAdjustmentTimer(int itemId, int itemPrice)
+	private void scheduleAdjustmentTimer(int itemId, long itemPrice)
 	{
 		long delay = AdjustmentTimerUtils.INITIAL_CHECK_DELAY_MS;
 		long deadline = System.currentTimeMillis() + delay;
@@ -1383,7 +1383,7 @@ public class AutoRecommendService
 	/**
 	 * Reset the adjustment timer for an item (e.g., after a partial fill).
 	 */
-	public synchronized void resetAdjustmentTimer(int itemId, int itemPrice)
+	public synchronized void resetAdjustmentTimer(int itemId, long itemPrice)
 	{
 		if (!active || !adjustments.hasBuyDeadline(itemId))
 		{
@@ -1498,7 +1498,7 @@ public class AutoRecommendService
 		String itemName = offer.getItemName();
 
 		// Get buy price (cost basis) for breakeven calculation
-		Integer costBasis = adjustments.getBuyPriceOrDefault(itemId, offer.getPrice());
+		Long costBasis = adjustments.getBuyPriceOrDefault(itemId, offer.getPrice());
 
 		log.debug("Auto-recommend: Checking buy adjustment for {} (price={}, {}min elapsed)",
 			itemName, offer.getPrice(), minutesSince);
@@ -1675,7 +1675,7 @@ public class AutoRecommendService
 
 	private void scheduleMissingSellTimer(OfferRecord offer, long now)
 	{
-		Integer buyPrice = adjustments.getBuyPriceOrDefault(offer.getItemId(), offer.getPrice());
+		Long buyPrice = adjustments.getBuyPriceOrDefault(offer.getItemId(), offer.getPrice());
 		long offerAgeMs = now - offer.getEffectiveLastActivityAtMillis();
 		long deadline = loginCheckDeadlineMs(offerAgeMs, now);
 		SellAdjustmentState state = new SellAdjustmentState(
@@ -1743,11 +1743,11 @@ public class AutoRecommendService
 
 	private void renderStaleOfferPrompt(OfferRecord offer)
 	{
-		Integer resellPrice = staleOffers.getResellPrice(offer.getItemId());
+		Long resellPrice = staleOffers.getResellPrice(offer.getItemId());
 		String overlayMsg;
 		if (resellPrice != null)
 		{
-			Integer net = staleOffers.getResellNet(offer.getItemId());
+			Long net = staleOffers.getResellNet(offer.getItemId());
 			String netSuffix = net == null ? ""
 				: String.format(" (%s%s)", net >= 0 ? "+" : "-", GpUtils.formatGP(Math.abs(net)));
 			// AC2 exit → cancel & re-sell; a priced buy prompt is the competitive buy
@@ -2011,7 +2011,7 @@ public class AutoRecommendService
 	 * uses the advised price. Idempotent — addToStaleQueue dedupes by item, and the price
 	 * map is refreshed on each poll.
 	 */
-	public synchronized void surfaceAdvisorResell(OfferRecord offer, int newPrice, Integer netProfitEstimate)
+	public synchronized void surfaceAdvisorResell(OfferRecord offer, long newPrice, Long netProfitEstimate)
 	{
 		surfaceAdvisorPrompt(offer, newPrice, netProfitEstimate, false);
 	}
@@ -2022,7 +2022,7 @@ public class AutoRecommendService
 	 * deliberately skips the session recommended price so the listing takes the no-offset
 	 * path. The statement order here reproduces both original methods exactly.
 	 */
-	private void surfaceAdvisorPrompt(OfferRecord offer, int price, Integer netProfitEstimate, boolean exit)
+	private void surfaceAdvisorPrompt(OfferRecord offer, long price, Long netProfitEstimate, boolean exit)
 	{
 		if (offer == null)
 		{
@@ -2061,7 +2061,7 @@ public class AutoRecommendService
 	 * an exit so the prompt reads "Cancel & re-sell" rather than "Adjust buy". Deliberately does
 	 * NOT set the session recommended price — that would flow through the offset-applying focus.
 	 */
-	public synchronized void surfaceAdvisorExitResell(OfferRecord offer, int resellPrice, Integer netProfitEstimate)
+	public synchronized void surfaceAdvisorExitResell(OfferRecord offer, long resellPrice, Long netProfitEstimate)
 	{
 		surfaceAdvisorPrompt(offer, resellPrice, netProfitEstimate, true);
 	}
@@ -2107,7 +2107,7 @@ public class AutoRecommendService
 		}
 
 		// Get the buy price (cost basis) — stored when the buy was placed
-		Integer buyPrice = adjustments.getBuyPrice(itemId);
+		Long buyPrice = adjustments.getBuyPrice(itemId);
 		if (buyPrice == null)
 		{
 			// Fallback: try recommendation queue
@@ -2291,7 +2291,7 @@ public class AutoRecommendService
 
 		if (response.isReadjustSell() && response.getRecommendedPrice() != null)
 		{
-			int newPrice = response.getRecommendedPrice();
+			long newPrice = response.getRecommendedPrice();
 			state.adjustmentCount++;
 
 			log.debug("Auto-recommend: Sell adjustment for {} — {} → {} gp (adj#{})",
@@ -2348,7 +2348,7 @@ public class AutoRecommendService
 			}
 
 			// Get buy price from stored prices or fall back to sell price
-			Integer buyPrice = adjustments.getBuyPrice(offer.getItemId());
+			Long buyPrice = adjustments.getBuyPrice(offer.getItemId());
 			if (buyPrice == null)
 			{
 				buyPrice = offer.getPrice();
@@ -2623,7 +2623,7 @@ public class AutoRecommendService
 	 */
 	private void focusSellForItem(int itemId, String itemName, int quantity)
 	{
-		Integer sellPrice = resolveBestSellPrice(itemId);
+		Long sellPrice = resolveBestSellPrice(itemId);
 
 		if (sellPrice == null || sellPrice <= 0)
 		{
@@ -2665,7 +2665,7 @@ public class AutoRecommendService
 		int sellableItemId = findNextSellableCollectedItem();
 		if (sellableItemId >= 0)
 		{
-			Integer sellPrice = resolveBestSellPrice(sellableItemId);
+			Long sellPrice = resolveBestSellPrice(sellableItemId);
 
 			if (sellPrice != null && sellPrice > 0)
 			{
@@ -2806,7 +2806,7 @@ public class AutoRecommendService
 
 	private boolean hasSellPrice(int itemId)
 	{
-		Integer price = resolveBestSellPrice(itemId);
+		Long price = resolveBestSellPrice(itemId);
 		return price != null && price > 0;
 	}
 
@@ -2884,10 +2884,10 @@ public class AutoRecommendService
 	 * placement, survives queue cycling), falling back to the live queue for a freshly-recommended
 	 * item not yet placed. Null when unknown (advisor then falls back to its existing behavior).
 	 */
-	public synchronized Integer getOriginalMargin(int itemId)
+	public synchronized Long getOriginalMargin(int itemId)
 	{
 		PlayerSession sess = plugin.getSession();
-		Integer persisted = sess == null ? null : sess.getOriginalMargin(itemId);
+		Long persisted = sess == null ? null : sess.getOriginalMargin(itemId);
 		if (persisted != null)
 		{
 			return persisted;
@@ -2902,17 +2902,17 @@ public class AutoRecommendService
 	 * adjustment) and is what the Flip Assist prompt reads. The panel's displayed
 	 * price is only a fallback for items the session hasn't seen yet.
 	 */
-	private Integer resolveBestSellPrice(int itemId)
+	private Long resolveBestSellPrice(int itemId)
 	{
-		Integer sessionPrice = plugin.getSession().getRecommendedPrice(itemId);
+		Long sessionPrice = plugin.getSession().getRecommendedPrice(itemId);
 		if (sessionPrice != null && sessionPrice > 0)
 		{
 			return sessionPrice;
 		}
-		IntFunction<Integer> provider = displayedSellPriceProvider;
+		IntFunction<Long> provider = displayedSellPriceProvider;
 		if (provider != null)
 		{
-			Integer smartPrice = provider.apply(itemId);
+			Long smartPrice = provider.apply(itemId);
 			if (smartPrice != null && smartPrice > 0)
 			{
 				return smartPrice;
@@ -3027,7 +3027,7 @@ public class AutoRecommendService
 	 * Create a buy FocusedFlip, invoke the focus callback, and update status text.
 	 * Centralizes the repeated pattern of showing a buy overlay with a status message.
 	 */
-	private void focusBuyOverlay(int itemId, String itemName, int buyPrice, int quantity, int sellPrice, String statusMsg)
+	private void focusBuyOverlay(int itemId, String itemName, long buyPrice, int quantity, long sellPrice, String statusMsg)
 	{
 		if (exitSuppressesBuys())
 		{
@@ -3235,7 +3235,7 @@ public class AutoRecommendService
 				{
 					continue;
 				}
-				Integer resolvedPrice = resolveBestSellPrice(itemId);
+				Long resolvedPrice = resolveBestSellPrice(itemId);
 				if (resolvedPrice == null || resolvedPrice <= 0)
 				{
 					// The sell price hasn't resolved yet (e.g. a transient wiki-price timeout).
